refactor: move removal logic to separate package
