+++ pkg/analyze/file_test.go
@@ -4,7 +4,6 @@ import (
 	"testing"
 	"time"
 
-	"github.com/dundee/gdu/v5/internal/testdir"
 	"github.com/dundee/gdu/v5/pkg/fs"
 	"github.com/stretchr/testify/assert"
 )
@@ -199,125 +198,6 @@ func TestRemoveByNameNotInDir(t *testing.T) {
 	dir.Files = dir.Files.RemoveByName("zzz")
 
 	assert.Equal(t, 1, len(dir.Files))
-}
-
-func TestRemoveFile(t *testing.T) {
-	dir := &Dir{
-		File: &File{
-			Name:  "xxx",
-			Size:  5,
-			Usage: 12,
-		},
-		ItemCount: 3,
-		BasePath:  ".",
-	}
-
-	subdir := &Dir{
-		File: &File{
-			Name:   "yyy",
-			Size:   4,
-			Usage:  8,
-			Parent: dir,
-		},
-		ItemCount: 2,
-	}
-	file := &File{
-		Name:   "zzz",
-		Size:   3,
-		Usage:  4,
-		Parent: subdir,
-	}
-	dir.Files = fs.Files{subdir}
-	subdir.Files = fs.Files{file}
-
-	err := RemoveItemFromDir(subdir, file)
-	assert.Nil(t, err)
-
-	assert.Equal(t, 0, len(subdir.Files))
-	assert.Equal(t, 1, subdir.ItemCount)
-	assert.Equal(t, int64(1), subdir.Size)
-	assert.Equal(t, int64(4), subdir.Usage)
-	assert.Equal(t, 1, len(dir.Files))
-	assert.Equal(t, 2, dir.ItemCount)
-	assert.Equal(t, int64(2), dir.Size)
-}
-
-func TestTruncateFile(t *testing.T) {
-	fin := testdir.CreateTestDir()
-	defer fin()
-
-	dir := &Dir{
-		File: &File{
-			Name:  "test_dir",
-			Size:  5,
-			Usage: 12,
-		},
-		ItemCount: 3,
-		BasePath:  ".",
-	}
-
-	subdir := &Dir{
-		File: &File{
-			Name:   "nested",
-			Size:   4,
-			Usage:  8,
-			Parent: dir,
-		},
-		ItemCount: 2,
-	}
-	file := &File{
-		Name:   "file2",
-		Size:   3,
-		Usage:  4,
-		Parent: subdir,
-	}
-	dir.Files = fs.Files{subdir}
-	subdir.Files = fs.Files{file}
-
-	err := EmptyFileFromDir(subdir, file)
-
-	assert.Nil(t, err)
-	assert.Equal(t, 1, len(subdir.Files))
-	assert.Equal(t, 2, subdir.ItemCount)
-	assert.Equal(t, int64(1), subdir.Size)
-	assert.Equal(t, int64(4), subdir.Usage)
-	assert.Equal(t, 1, len(dir.Files))
-	assert.Equal(t, 3, dir.ItemCount)
-	assert.Equal(t, int64(2), dir.Size)
-}
-
-func TestTruncateFileWithErr(t *testing.T) {
-	dir := &Dir{
-		File: &File{
-			Name:  "xxx",
-			Size:  5,
-			Usage: 12,
-		},
-		ItemCount: 3,
-		BasePath:  ".",
-	}
-
-	subdir := &Dir{
-		File: &File{
-			Name:   "yyy",
-			Size:   4,
-			Usage:  8,
-			Parent: dir,
-		},
-		ItemCount: 2,
-	}
-	file := &File{
-		Name:   "zzz",
-		Size:   3,
-		Usage:  4,
-		Parent: subdir,
-	}
-	dir.Files = fs.Files{subdir}
-	subdir.Files = fs.Files{file}
-
-	err := EmptyFileFromDir(subdir, file)
-
-	assert.Contains(t, err.Error(), "no such file or directory")
 }
 
 func TestUpdateStats(t *testing.T) {
